Extract duplicate HTTPClientSpy helper into a shared scope
